+++ src/main/java/demesup/servlet/PictureServlet.java
@@ -1,6 +1,5 @@
-package servlet;
+package demesup.servlet;
 
-import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
