+++ src/main/resources/hibernate.cfg.xml
@@ -14,7 +14,7 @@
         <property name="connection.pool_size">10</property>
         <property name="hbm2ddl.auto">update</property>
 
-        <mapping class="model.Activity"/>
+        <mapping class="demesup.model.Activity"/>
 
     </session-factory>
 </hibernate-configuration>
